Refactor: root layout 수정
