+++ src/components/layout/Header.tsx
@@ -23,6 +23,7 @@ const Header = () => {
       setSearchData("");
     }
   };
+
   return (
     <HeaderContent>
       <Link to="/">
