+++ src/components/layout/Layout.tsx
@@ -1,22 +1,38 @@
-import Footer from "components/Footer";
-import Header from "components/Header";
-import { Outlet } from "react-router-dom";
+import Footer from "components/layout/Footer";
+import Header from "components/layout/Header";
+import PageHeader from "components/layout/PageHeader";
+import { Outlet, useLocation } from "react-router-dom";
 import styled from "styled-components";
 
 const Layout = () => {
+  const { pathname } = useLocation();
+
+  console.log(pathname.includes("/"));
+
   return (
     <>
-      <Header />
+      {pathname.includes("mypage") ||
+      pathname.includes("signup") ||
+      pathname.includes("signin") ||
+      pathname.includes("cart") ? (
+        <PageHeader />
+      ) : (
+        <Header />
+      )}
       <Main>
         <Outlet />
       </Main>
-      <Footer />
+      {pathname.includes("signup") ||
+      pathname.includes("signin") ||
+      pathname.includes("cart") ? null : (
+        <Footer />
+      )}
     </>
   );
 };
 
 const Main = styled.main`
-  padding: 90px 12px 30px;
+  padding: 0 12px 30px;
 `;
 
 export default Layout;
