+++ src/router/Router.tsx
@@ -17,32 +17,28 @@ import NotFound from "pages/NotFound";
 import Order from "pages/Order";
 import User from "pages/User";
 import CompleteOrder from "pages/CompleteOrder";
-import MyPageLayout from "components/layout/MypageLayout";
-import GotoBackLayout from "components/layout/GotoBackLayout";
 import Password from "pages/Password";
 
 const Router = createBrowserRouter(
   createRoutesFromElements(
     <Route>
-      <Route element={<GotoBackLayout />}>
+      <Route element={<Root />}>
         <Route path="/signin" element={<SignIn />} />
         <Route path="/signup" element={<SingUp />} />
-      </Route>
-
-      <Route path="/cart" element={<Cart />} />
-      <Route path="/product/:id" element={<ProductDetail />} />
-      <Route path="/completeOrder" element={<CompleteOrder />} />
-      <Route element={<Root />}>
+        <Route path="/cart" element={<Cart />} />
+        <Route path="/product/:id" element={<ProductDetail />} />
+        <Route path="/completeOrder" element={<CompleteOrder />} />
         <Route path="/" element={<Main />} />
         <Route path="/search/:value" element={<Search />} />
+        <Route path="/mypage">
+          <Route path="/mypage" element={<MyPage />} />
+          <Route path="user" element={<User />} />
+          <Route path="user/password" element={<Password />} />
+          <Route path="order" element={<Order />} />
+          <Route path="likes" element={<Likes />} />
+        </Route>
       </Route>
-      <Route path="/mypage" element={<MyPageLayout />}>
-        <Route path="/mypage" element={<MyPage />} />
-        <Route path="user" element={<User />} />
-        <Route path="user/password" element={<Password />} />
-        <Route path="order" element={<Order />} />
-        <Route path="likes" element={<Likes />} />
-      </Route>
+
       <Route path="*" element={<NotFound />} />
     </Route>,
   ),
